Basic prow job for Kubetest2 GCE deployer
Will be run manually while the test script is tuned, then will be
turned on as a presubmit for all changes that affect the GCE deployer.
The cloud-provider-gcp repo is needed as an extra dependency because
we need a repo to build that also contains the cluster/gce scripts
like kube-up.sh. k/k could also be used, but cloud-provider-gcp is the
expected and encouraged future default.